+++ cmd/workflow-engine/cli/root.go
@@ -15,6 +15,7 @@ var (
 
 func NewWorkflowEngineCommand() *cobra.Command {
 	viper.SetConfigName("workflow-engine")
+	pipelines.BindEnvs(viper.GetViper())
 	pipelines.SetDefaults(viper.GetViper())
 
 	versionCmd := newBasicCommand("version", "print version information", runVersion)