+++ pkg/shell/docker.go
@@ -1,6 +1,8 @@
 package shell
 
-import "os/exec"
+import (
+	"os/exec"
+)
 
 type DockerAlias int8
 
@@ -28,4 +30,113 @@ func DockerSave(optionFuncs ...OptionFunc) ExitCode {
 	default:
 		return ExitBadConfiguration
 	}
+}
+
+// DockerBuild a container image, supports CLI aliases to docker build (ex. podman build)
+//
+// Requirements: WithImageBuildOptions, optional WithDockerAlias
+//
+// Outputs: debug to STDERR
+func DockerBuild(optionFuncs ...OptionFunc) ExitCode {
+	o := newOptions(optionFuncs...)
+	// this parses argument values to determine the flags
+	args := o.imageBuildOptions.args()
+	switch o.dockerAlias {
+	case DockerAliasDocker:
+		cmd := exec.Command("docker", args...)
+		return run(cmd, o)
+	case DockerAliasPodman:
+		cmd := exec.Command("podman", args...)
+		return run(cmd, o)
+	default:
+		return ExitBadConfiguration
+	}
+}
+
+// DockerPush a container image, supports CLI aliases to docker build (ex. podman build)
+//
+// Requirements: WithImageName
+//
+// Outputs: debug to STDERR
+func DockerPush(optionFuncs ...OptionFunc) ExitCode {
+	o := newOptions(optionFuncs...)
+	// this parses argument values to determine the flags
+	switch o.dockerAlias {
+	case DockerAliasDocker:
+		cmd := exec.Command("docker", "push", o.imageName)
+		return run(cmd, o)
+	case DockerAliasPodman:
+		cmd := exec.Command("podman", "push", o.imageName)
+		return run(cmd, o)
+	default:
+		return ExitBadConfiguration
+	}
+}
+
+// DockerInfo print system configuration information
+//
+// Requirements: optional WithDockerAlias
+//
+// Outputs: debug to STDERR
+func DockerInfo(optionFuncs ...OptionFunc) ExitCode {
+	o := newOptions(optionFuncs...)
+	// this parses argument values to determine the flags
+	switch o.dockerAlias {
+	case DockerAliasDocker:
+		cmd := exec.Command("docker", "info")
+		return run(cmd, o)
+	case DockerAliasPodman:
+		cmd := exec.Command("podman", "info")
+		return run(cmd, o)
+	default:
+		return ExitBadConfiguration
+	}
+}
+
+// ImageBuildOptions are specific to Docker builds
+type ImageBuildOptions struct {
+	ImageName    string
+	BuildDir     string
+	Dockerfile   string
+	Target       string
+	Platform     string
+	SquashLayers bool
+	CacheTo      string
+	CacheFrom    string
+	buildArgs    []string
+}
+
+func (o ImageBuildOptions) args() []string {
+	args := []string{"build"}
+
+	flags := map[string]string{
+		"--file":       o.Dockerfile,
+		"--target":     o.Target,
+		"--platform":   o.Platform,
+		"--cache-to":   o.CacheTo,
+		"--cache-from": o.CacheFrom,
+	}
+
+	// append to the args list if not ""
+	for flag, value := range flags {
+		if value != "" {
+			args = append(args, flag, value)
+		}
+	}
+
+	// Special case for build arguments
+	for _, arg := range o.buildArgs {
+		if arg != "" {
+			args = append(args, arg)
+		}
+	}
+
+	// Squash is a bool flag
+	if o.SquashLayers {
+		args = append(args, "--squash-all")
+	}
+
+	args = append(args, o.BuildDir)
+
+	return args
 }